Finish moving reusable/shared functionality into the new lib directory.
