+++ src/lib/othent-kms-client/client.types.ts
@@ -1,0 +1,3 @@
+export interface CommonEncodedRequestData {
+  encodedData: string;
+}
