+++ src/lib/othent-kms-client/operations/server-info.ts
@@ -1,8 +1,8 @@
 import { AxiosInstance } from "axios";
-import { CommonEncodedRequestData } from "./common.types";
 import { parseErrorResponse } from "../../utils/errors/error.utils";
 import { OthentAuth0Client } from "../../auth/auth0";
-import { Route } from "./common.constants";
+import { Route } from "../client.constants";
+import { CommonEncodedRequestData } from "../client.types";
 
 export interface ServerInfoOptions {
   includeToken?: boolean;
